requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==42.0.5
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
jwt==1.3.1
pycparser==2.22
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing-extensions==4.11.0
